+++ misago/settings.py
@@ -76,6 +76,7 @@ INSTALLED_APPS = [
     "misago.search",
     "misago.socialauth",
     "misago.threads",
+    "misago.threadupdates",
     "misago.faker",
     "misago.graphql",
     "misago.apiv2",
