+++ misago/threads/enums.py
@@ -1,4 +1,4 @@
-from enum import IntEnum, StrEnum
+from enum import IntEnum
 
 from django.utils.translation import pgettext_lazy
 
@@ -35,33 +35,3 @@ class ThreadsListsPolling(IntEnum):
                 pgettext_lazy("category lists polling choice", "Disable"),
             ),
         )
-
-
-class ThreadUpdateActionName(StrEnum):
-    APPROVED = "approved"
-
-    PINNED_GLOBALLY = "pinned_globally"
-    PINNED_IN_CATEGORY = "pinned_in_category"
-    UNPINNED = "unpinned"
-
-    LOCKED = "locked"
-    OPENED = "opened"
-
-    MOVED = "moved"
-
-    MERGED = "merged"
-    SPLIT = "split"
-
-    HID = "hid"
-    UNHID = "unhid"
-
-    CHANGED_TITLE = "changed_title"
-
-    TOOK_OWNERSHIP = "took_ownership"
-    CHANGED_OWNER = "changed_owner"
-
-    JOINED = "joined"
-    LEFT = "left"
-
-    INVITED_PARTICIPANT = "invited_participant"
-    REMOVED_PARTICIPANT = "removed_participant"
