+++ misago/threads/models/__init__.py
@@ -5,7 +5,6 @@ from .postedit import PostEdit
 from .postlike import PostLike
 from .thread import Thread
 from .threadparticipant import ThreadParticipant
-from .threadupdate import ThreadUpdate
 from .poll import Poll
 from .pollvote import PollVote
 
@@ -17,7 +16,6 @@ __all__ = [
     "PostLike",
     "Thread",
     "ThreadParticipant",
-    "ThreadUpdate",
     "Poll",
     "PollVote",
 ]
