+++ misago/threads/tests/test_delete_thread_update_view.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 from ...permissions.enums import CategoryPermission
 from ...permissions.models import CategoryGroupPermission, Moderator
 from ...test import assert_contains
-from ..models import ThreadUpdate
+from ...threadupdates.models import ThreadUpdate
 
 
 def test_delete_thread_update_view_returns_404_error_for_not_found_thread(user_client):
