+++ misago/threads/views/replies.py
@@ -34,6 +34,7 @@ from ...readtracker.tracker import (
 )
 from ...readtracker.privatethreads import unread_private_threads_exist
 from ...readtracker.threads import is_category_read
+from ...threadupdates.models import ThreadUpdate
 from ..hooks import (
     get_private_thread_replies_page_context_data_hook,
     get_private_thread_replies_page_posts_queryset_hook,
@@ -42,7 +43,7 @@ from ..hooks import (
     get_thread_replies_page_posts_queryset_hook,
     get_thread_replies_page_thread_queryset_hook,
 )
-from ..models import Post, Thread, ThreadUpdate
+from ..models import Post, Thread
 from ..paginator import ThreadRepliesPage
 from .generic import PrivateThreadView, ThreadView
