+++ misago/threadupdates/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class MisagoThreadUpdatesConfig(AppConfig):
+    name = "misago.threadupdates"
+    label = "misago_threadupdates"
+    verbose_name = "Misago Thread Updates"
