+++ misago/threadupdates/hooks/__init__.py
@@ -1,0 +1,11 @@
+from .create_thread_update import create_thread_update_hook
+from .delete_thread_update import delete_thread_update_hook
+from .hide_thread_update import hide_thread_update_hook
+from .unhide_thread_update import unhide_thread_update_hook
+
+__all__ = [
+    "create_thread_update_hook",
+    "delete_thread_update_hook",
+    "hide_thread_update_hook",
+    "unhide_thread_update_hook",
+]
